split main in modules
