+++ Makefile
@@ -72,7 +72,7 @@ pytest:
 
 
 doctest:
-	MATHICS_CHARACTER_ENCODING="ASCII" $(PYTHON) -m mathics.docpipeline -l pymathics.natlang -c  "Natural Language Processing"  $o
+	MATHICS_CHARACTER_ENCODING="ASCII" $(PYTHON) -m mathics.docpipeline -l pymathics.natlang -c 'Natural Language Processing' $o
 
 
 # #: Make Mathics PDF manual
@@ -89,5 +89,4 @@ ChangeLog: rmChangeLog
 
 #: Run pytest consistency and style checks
 check-consistency-and-style:
-	# MATHICS_LINT=t $(PYTHON) -m pytest test/consistency-and-style
-	echo "check-consistency-and-style deactivated. Activate me later. "
+	MATHICS_LINT=t $(PYTHON) -m pytest test/consistency-and-style
